+++ BTCPayServer.Tests/LightningTests.cs
@@ -8,7 +8,6 @@ using BTCPayServer.Client.Models;
 using BTCPayServer.Controllers;
 using BTCPayServer.Events;
 using BTCPayServer.Lightning;
-using BTCPayServer.Lightning.Charge;
 using BTCPayServer.Models.StoreViewModels;
 using BTCPayServer.Payments;
 using BTCPayServer.Payments.Lightning;
@@ -22,6 +21,7 @@ using Xunit;
 using static Microsoft.Playwright.Assertions;
 using CreateInvoiceRequest = BTCPayServer.Client.Models.CreateInvoiceRequest;
 using BTCPayServer.Data;
+using BTCPayServer.Lightning.CLightning;
 using LightningAddressData = BTCPayServer.Client.Models.LightningAddressData;
 
 namespace BTCPayServer.Tests;
@@ -40,7 +40,7 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         await tester.EnsureChannelsSetup();
         var user = tester.NewAccount();
         await user.GrantAccessAsync(true);
-        user.RegisterLightningNode("BTC", LightningConnectionType.CLightning);
+        user.RegisterLightningNode("BTC", LightningTestImplementation.CoreLightning);
 
         var client = await user.CreateClient(Policies.Unrestricted);
         var invoices = new Task<Client.Models.InvoiceData>[5];
@@ -103,11 +103,11 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         await tester.EnsureChannelsSetup();
         var user = tester.NewAccount();
         await user.GrantAccessAsync(true);
-        user.RegisterLightningNode("BTC", LightningConnectionType.CLightning, false);
+        user.RegisterLightningNode("BTC", LightningTestImplementation.CoreLightning, false);
 
         var merchant = tester.NewAccount();
         await merchant.GrantAccessAsync(true);
-        merchant.RegisterLightningNode("BTC", LightningConnectionType.LndREST);
+        merchant.RegisterLightningNode("BTC", LightningTestImplementation.LND);
         var merchantClient = await merchant.CreateClient($"{Policies.CanUseLightningNodeInStore}:{merchant.StoreId}");
         var merchantInvoice = await merchantClient.CreateLightningInvoice(merchant.StoreId, "BTC",
             new CreateLightningInvoiceRequest(LightMoney.Satoshis(1_000), "hey", TimeSpan.FromSeconds(60)));
@@ -272,7 +272,7 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         var user = tester.NewAccount();
         await user.GrantAccessAsync(true);
 
-        var types = new[] { LightningConnectionType.LndREST, LightningConnectionType.CLightning };
+        var types = new[] { LightningTestImplementation.LND, LightningTestImplementation.CoreLightning };
         foreach (var type in types)
         {
             user.RegisterLightningNode("BTC", type);
@@ -614,7 +614,7 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
     [Trait("Lightning", "Lightning")]
     public async Task CanSendLightningPaymentCLightning()
     {
-        await ProcessLightningPayment(LightningConnectionType.CLightning);
+        await ProcessLightningPayment(LightningTestImplementation.CoreLightning);
     }
 
     [Fact(Timeout = 60 * 2 * 1000)]
@@ -622,10 +622,10 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
     [Trait("Lightning", "Lightning")]
     public async Task CanSendLightningPaymentLnd()
     {
-        await ProcessLightningPayment(LightningConnectionType.LndREST);
+        await ProcessLightningPayment(LightningTestImplementation.LND);
     }
 
-    async Task ProcessLightningPayment(string type)
+    async Task ProcessLightningPayment(LightningTestImplementation type)
     {
         // For easier debugging and testing
         // LightningLikePaymentHandler.LIGHTNING_TIMEOUT = int.MaxValue;
@@ -681,7 +681,7 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         var user = tester.NewAccount();
         await user.GrantAccessAsync(true);
         await user.RegisterDerivationSchemeAsync("BTC");
-        await user.RegisterLightningNodeAsync("BTC", LightningConnectionType.CLightning);
+        await user.RegisterLightningNodeAsync("BTC", LightningTestImplementation.CoreLightning);
         await user.SetNetworkFeeMode(NetworkFeeMode.Never);
         await user.ModifyGeneralSettings(p => p.SpeedPolicy = SpeedPolicy.HighSpeed);
         var invoice = await user.BitPay.CreateInvoiceAsync(new Invoice(0.0001m, "BTC"));
@@ -748,9 +748,10 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         Assert.IsType<ViewResult>(storeResponse);
         Assert.IsType<ViewResult>(storeController.SetupLightningNode(user.StoreId, "BTC"));
 
+        var address = ((CLightningClient)tester.CustomerLightningD).Address.AbsoluteUri;
         await storeController.SetupLightningNode(user.StoreId, new LightningNodeViewModel
         {
-            ConnectionString = $"type=charge;server={tester.MerchantCharge.Client.Uri.AbsoluteUri};allowinsecure=true",
+            ConnectionString = $"type=clightning;server={address}",
             SkipPortTest = true // We can't test this as the IP can't be resolved by the test host :(
         }, "test", "BTC");
         Assert.False(storeController.TempData.ContainsKey(WellKnownTempData.ErrorMessage));
@@ -760,12 +761,12 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         Assert.IsType<RedirectToActionResult>(await storeController.SetupLightningNode(user.StoreId,
             new LightningNodeViewModel
             {
-                ConnectionString = $"type=charge;server={tester.MerchantCharge.Client.Uri.AbsoluteUri};allowinsecure=true"
+                ConnectionString = $"type=clightning;server={address}"
             }, "save", "BTC"));
 
         // Make sure old connection string format does not work
         Assert.IsType<RedirectToActionResult>(await storeController.SetupLightningNode(user.StoreId,
-            new LightningNodeViewModel { ConnectionString = tester.MerchantCharge.Client.Uri.AbsoluteUri },
+            new LightningNodeViewModel { ConnectionString = address },
             "save", "BTC"));
 
         storeResponse = storeController.LightningSettings(user.StoreId, "BTC");
@@ -837,7 +838,7 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
     public async Task CanDoLightningInternalNodeMigration()
     {
         using var tester = CreateServerTester(newDb: true);
-        tester.ActivateLightning(LightningConnectionType.CLightning);
+        tester.ActivateLightning(LightningTestImplementation.CoreLightning);
         await tester.StartAsync();
         var acc = tester.NewAccount();
         await acc.GrantAccessAsync(true);
@@ -864,7 +865,7 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         Assert.Equal(derivation, v.AccountOriginal);
         Assert.Equal(xpub, v.GetFirstAccountKeySettings().AccountKey.ToString());
 
-        await acc.RegisterLightningNodeAsync("BTC", LightningConnectionType.CLightning);
+        await acc.RegisterLightningNodeAsync("BTC", LightningTestImplementation.CoreLightning);
         store = await tester.PayTester.StoreRepository.FindStore(acc.StoreId);
 
         pmi = PaymentTypes.LN.GetPaymentMethodId("BTC");
@@ -890,35 +891,6 @@ public class LightningTests(ITestOutputHelper testOutputHelper) : UnitTestBase(t
         Assert.Null(conf["CryptoCode"]); // Osolete
         Assert.Null(conf["connectionString"]); // Null, so should be stripped
         Assert.Null(conf["DisableBOLT11PaymentOption"]); // Old garbage cleaned
-
-        // Test if legacy lightning charge settings are converted to LightningConnectionString
-        store.DerivationStrategies = new JObject()
-        {
-            new JProperty("BTC_LightningLike", new JObject()
-            {
-                new JProperty("LightningChargeUrl", "http://mycharge.com/"),
-                new JProperty("Username", "usr"),
-                new JProperty("Password", "pass"),
-                new JProperty("CryptoCode", "BTC"),
-                new JProperty("PaymentId", "someshit"),
-            })
-        }.ToString();
-        await tester.PayTester.StoreRepository.UpdateStore(store);
-        await tester.RestartMigration();
-        store = await tester.PayTester.StoreRepository.FindStore(acc.StoreId);
-        Assert.NotNull(store);
-        lnMethod = store.GetPaymentMethodConfig<LightningPaymentMethodConfig>(pmi, handlers);
-        Assert.NotNull(lnMethod?.GetExternalLightningUrl());
-
-        var url = lnMethod.GetExternalLightningUrl();
-        LightningConnectionStringHelper.ExtractValues(url, out var connType);
-        Assert.Equal(LightningConnectionType.Charge, connType);
-        var client = Assert.IsType<ChargeClient>(tester.PayTester.GetService<LightningClientFactoryService>()
-            .Create(url, tester.NetworkProvider.GetNetwork<BTCPayNetwork>("BTC")));
-        var auth = Assert.IsType<ChargeAuthentication.UserPasswordAuthentication>(client.ChargeAuthentication);
-
-        Assert.Equal("pass", auth.NetworkCredential.Password);
-        Assert.Equal("usr", auth.NetworkCredential.UserName);
 
         // Test if lightning connection strings get migrated to internal
         store.DerivationStrategies = new JObject()